Refactor Assert.ContainsSingle.cs into focused files (#9948)
Split the 572-line Assert.ContainsSingle.cs into three focused partial-class files:
- Assert.ContainsSingle.cs (IEnumerable overloads + failure helpers)
- Assert.ContainsSingle.Span.cs (span/memory overloads)
- Assert.ContainsSingle.InterpolatedStringHandler.cs (AssertSingleInterpolatedStringHandler struct)

No public API changes; all ContainsSingle unit tests pass.

Copilot-Session: deab99dd-54e5-4459-af4b-ffc033638860
