+++ src/TestFramework/TestFramework/Assertions/Assert.ContainsSingle.cs
@@ -1,123 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.ComponentModel;
-
 namespace Microsoft.VisualStudio.TestTools.UnitTesting;
 
 public sealed partial class Assert
 {
-    /// <summary>
-    /// Provides an interpolated string handler used by <c>Assert.ContainsSingle</c> overloads
-    /// that only allocates and formats the message when the assertion is failing.
-    /// </summary>
-    /// <typeparam name="TItem">The type of item in the collection.</typeparam>
-    /// <remarks>
-    /// This type is intended to be used by the compiler; users should not reference it directly.
-    /// </remarks>
-    [StackTraceHidden]
-    [InterpolatedStringHandler]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    [GenerateAssertInterpolatedStringAppendMethods]
-    public readonly partial struct AssertSingleInterpolatedStringHandler<TItem>
-    {
-        private readonly StringBuilder? _builder;
-        private readonly int _actualCount;
-        private readonly TItem? _item;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertSingleInterpolatedStringHandler{TItem}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="collection">The collection being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertSingleInterpolatedStringHandler(int literalLength, int formattedCount, IEnumerable<TItem> collection, out bool shouldAppend)
-        {
-            _actualCount = collection.Count();
-            shouldAppend = _actualCount != 1;
-            if (shouldAppend)
-            {
-                _builder = new StringBuilder(literalLength + formattedCount);
-            }
-            else
-            {
-                _item = collection.First();
-            }
-        }
-
-#if NETCOREAPP3_1_OR_GREATER
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertSingleInterpolatedStringHandler{TItem}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="collection">The collection being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertSingleInterpolatedStringHandler(int literalLength, int formattedCount, ReadOnlySpan<TItem> collection, out bool shouldAppend)
-        {
-            _actualCount = collection.Length;
-            shouldAppend = _actualCount != 1;
-            if (shouldAppend)
-            {
-                _builder = new StringBuilder(literalLength + formattedCount);
-            }
-            else
-            {
-                _item = collection[0];
-            }
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertSingleInterpolatedStringHandler{TItem}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="collection">The collection being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertSingleInterpolatedStringHandler(int literalLength, int formattedCount, Span<TItem> collection, out bool shouldAppend)
-            : this(literalLength, formattedCount, (ReadOnlySpan<TItem>)collection, out shouldAppend)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertSingleInterpolatedStringHandler{TItem}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="collection">The collection being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertSingleInterpolatedStringHandler(int literalLength, int formattedCount, ReadOnlyMemory<TItem> collection, out bool shouldAppend)
-            : this(literalLength, formattedCount, collection.Span, out shouldAppend)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AssertSingleInterpolatedStringHandler{TItem}"/> struct.
-        /// </summary>
-        /// <param name="literalLength">The number of constant characters in the interpolated string.</param>
-        /// <param name="formattedCount">The number of interpolation expressions in the interpolated string.</param>
-        /// <param name="collection">The collection being asserted; the message is only computed when the assertion fails.</param>
-        /// <param name="shouldAppend">When this method returns, indicates whether the interpolated string should be evaluated.</param>
-        public AssertSingleInterpolatedStringHandler(int literalLength, int formattedCount, Memory<TItem> collection, out bool shouldAppend)
-            : this(literalLength, formattedCount, (ReadOnlyMemory<TItem>)collection, out shouldAppend)
-        {
-        }
-
-#endif
-
-        internal TItem ComputeAssertion(string collectionExpression)
-        {
-            if (_builder is not null)
-            {
-                ReportAssertContainsSingleFailed(_actualCount, _builder.ToString(), collectionExpression);
-            }
-
-            return _item!;
-        }
-    }
-
 #pragma warning disable RS0026 // Do not add multiple public overloads with optional parameters
 
     /// <summary>
@@ -301,246 +188,6 @@ public sealed partial class Assert
     }
 
     #endregion // ContainsSingle
-
-#if NETCOREAPP3_1_OR_GREATER
-
-    #region ContainsSingle span/memory
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-#pragma warning disable IDE0060 // Remove unused parameter
-    public static T ContainsSingle<T>(ReadOnlySpan<T> collection, [InterpolatedStringHandlerArgument(nameof(collection))] ref AssertSingleInterpolatedStringHandler<T> message, [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-#pragma warning restore IDE0060 // Remove unused parameter
-        => message.ComputeAssertion(collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-    public static T ContainsSingle<T>(ReadOnlySpan<T> collection, string? message = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-    {
-        if (collection.Length == 1)
-        {
-            return collection[0];
-        }
-
-        ReportAssertContainsSingleFailed(collection.Length, message, collectionExpression);
-        // Unreachable code but compiler cannot work it out
-        return default!;
-    }
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-#pragma warning disable IDE0060 // Remove unused parameter
-    public static T ContainsSingle<T>(Span<T> collection, [InterpolatedStringHandlerArgument(nameof(collection))] ref AssertSingleInterpolatedStringHandler<T> message, [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-#pragma warning restore IDE0060 // Remove unused parameter
-        => message.ComputeAssertion(collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-    public static T ContainsSingle<T>(Span<T> collection, string? message = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle((ReadOnlySpan<T>)collection, message, collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-#pragma warning disable IDE0060 // Remove unused parameter
-    public static T ContainsSingle<T>(ReadOnlyMemory<T> collection, [InterpolatedStringHandlerArgument(nameof(collection))] ref AssertSingleInterpolatedStringHandler<T> message, [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-#pragma warning restore IDE0060 // Remove unused parameter
-        => message.ComputeAssertion(collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-    public static T ContainsSingle<T>(ReadOnlyMemory<T> collection, string? message = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle(collection.Span, message, collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-#pragma warning disable IDE0060 // Remove unused parameter
-    public static T ContainsSingle<T>(Memory<T> collection, [InterpolatedStringHandlerArgument(nameof(collection))] ref AssertSingleInterpolatedStringHandler<T> message, [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-#pragma warning restore IDE0060 // Remove unused parameter
-        => message.ComputeAssertion(collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item.</returns>
-    public static T ContainsSingle<T>(Memory<T> collection, string? message = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle(collection.Span, message, collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element that matches the given predicate.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="predicate">A function to test each element for a condition.</param>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="predicateExpression">
-    /// The syntactic expression of predicate as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item that matches the predicate.</returns>
-    public static T ContainsSingle<T>(Func<T, bool> predicate, ReadOnlySpan<T> collection, string? message = "", [CallerArgumentExpression(nameof(predicate))] string predicateExpression = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-    {
-        TelemetryCollector.TrackAssertionCall("Assert.ContainsSingle");
-        T firstMatch = default!;
-        int matchCount = 0;
-        foreach (T item in collection)
-        {
-            if (!predicate(item))
-            {
-                continue;
-            }
-
-            if (matchCount == 0)
-            {
-                firstMatch = item;
-            }
-
-            matchCount++;
-        }
-
-        if (matchCount == 1)
-        {
-            return firstMatch;
-        }
-
-        ReportAssertSingleMatchFailed(matchCount, message, predicateExpression, collectionExpression);
-        // Unreachable code but compiler cannot work it out
-        return default!;
-    }
-
-    /// <summary>
-    /// Tests whether the specified span contains exactly one element that matches the given predicate.
-    /// </summary>
-    /// <typeparam name="T">The type of the span items.</typeparam>
-    /// <param name="predicate">A function to test each element for a condition.</param>
-    /// <param name="collection">The span.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="predicateExpression">
-    /// The syntactic expression of predicate as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item that matches the predicate.</returns>
-    public static T ContainsSingle<T>(Func<T, bool> predicate, Span<T> collection, string? message = "", [CallerArgumentExpression(nameof(predicate))] string predicateExpression = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle(predicate, (ReadOnlySpan<T>)collection, message, predicateExpression, collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element that matches the given predicate.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="predicate">A function to test each element for a condition.</param>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="predicateExpression">
-    /// The syntactic expression of predicate as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item that matches the predicate.</returns>
-    public static T ContainsSingle<T>(Func<T, bool> predicate, ReadOnlyMemory<T> collection, string? message = "", [CallerArgumentExpression(nameof(predicate))] string predicateExpression = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle(predicate, collection.Span, message, predicateExpression, collectionExpression);
-
-    /// <summary>
-    /// Tests whether the specified memory contains exactly one element that matches the given predicate.
-    /// </summary>
-    /// <typeparam name="T">The type of the memory items.</typeparam>
-    /// <param name="predicate">A function to test each element for a condition.</param>
-    /// <param name="collection">The memory.</param>
-    /// <param name="message">The message to display when the assertion fails.</param>
-    /// <param name="predicateExpression">
-    /// The syntactic expression of predicate as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <param name="collectionExpression">
-    /// The syntactic expression of collection as given by the compiler via caller argument expression.
-    /// Users shouldn't pass a value for this parameter.
-    /// </param>
-    /// <returns>The item that matches the predicate.</returns>
-    public static T ContainsSingle<T>(Func<T, bool> predicate, Memory<T> collection, string? message = "", [CallerArgumentExpression(nameof(predicate))] string predicateExpression = "", [CallerArgumentExpression(nameof(collection))] string collectionExpression = "")
-        => ContainsSingle(predicate, collection.Span, message, predicateExpression, collectionExpression);
-
-    #endregion // ContainsSingle span/memory
-
-#endif
 
     [DoesNotReturn]
     private static void ReportAssertContainsSingleFailed(int actualCount, string? userMessage, string collectionExpression)
